+++ backend/expenses/Touir.ExpensesManager.Expenses/Controllers/Requests/CreateExpenseRequest.cs
@@ -1,10 +1,10 @@
 namespace Touir.ExpensesManager.Expenses.Controllers.Requests
 {
-    public class CreateExpenseRequest
+    public class CreateExpenseRequest : IExpenseRequest
     {
-        public decimal Amount { get; set; }
-        public int CurrencyId { get; set; }
-        public DateOnly Date { get; set; }
+        public required decimal Amount { get; set; }
+        public required int CurrencyId { get; set; }
+        public required DateOnly Date { get; set; }
         public int? CategoryId { get; set; }
         public int? SubcategoryId { get; set; }
         public string? Description { get; set; }